Create everlasting-fairytale
This is everlasting-fairytale Power Bi Application. This power bi solution attempted to clarify and demonstrate the Hootan Dehbonei Company activity. It would enhance the productivity by showing the necessary information about  Sale, human resources, IT to maximize its operation.
We are one of the most experienced web designers and software engineering companies in the IT industry. We have designed and developed apps; Everlasting Fairytale, Everlasting Fairytale QuickStart etc